The last bug: The right and bottom border always appeared too small. That is due to rounding errors which make the diamonds move to the lower right. These rounding errors are nearly fixed, the outstanding difference is virtually invisible.
CCMAIL: eugene.trounev@gmail.com

svn path=/trunk/KDE/kdegames/kdiamond/; revision=791841

+++ src/diamond.h
@@ -47,7 +47,7 @@ class Diamond : public QObject, public QGraphicsPixmapItem
 {
     Q_OBJECT
     public:
-        Diamond(int xIndex, int yIndex, qreal xPos, qreal yPos, KDiamond::Color color, Board *board);
+        Diamond(int xIndex, int yIndex, int xPos, int yPos, KDiamond::Color color, Board *board);
 
         KDiamond::Color color() const;
         int xIndex() const;
